+++ src/BandcampDownloader/Core/UserSettings.cs
@@ -7,10 +7,23 @@ namespace BandcampDownloader {
     public enum Language {
         [Description("English")]
         en,
+        [Description("Deutsch (German)")]
+        de,
         [Description("Français (French)")]
         fr,
-        [Description("Deutsch (German)")]
-        de
+        [Description("Italiano (Italian)")]
+        it
+    }
+
+    public enum PlaylistFormat {
+        [Description("M3U")]
+        m3u,
+        [Description("PLS")]
+        pls,
+        [Description("WPL (Windows Media Player)")]
+        wpl,
+        [Description("ZPL (Zune Media Player)")]
+        zpl
     }
 
     public enum ProxyType {
@@ -65,6 +78,9 @@ namespace BandcampDownloader {
         Boolean CoverArtInTagsResize { get; set; }
 
         [Option(DefaultValue = false)]
+        Boolean CreatePlaylist { get; set; }
+
+        [Option(DefaultValue = false)]
         Boolean DownloadArtistDiscography { get; set; }
 
         [Option(DefaultValue = 7)]
@@ -94,7 +110,16 @@ namespace BandcampDownloader {
         Language Language { get; set; }
 
         [Option(DefaultValue = true)]
+        Boolean M3uExtended { get; set; }
+
+        [Option(DefaultValue = true)]
         Boolean ModifyTags { get; set; }
+
+        [Option(DefaultValue = "{album}")]
+        String PlaylistFileNameFormat { get; set; }
+
+        [Option(DefaultValue = PlaylistFormat.m3u)]
+        PlaylistFormat PlaylistFormat { get; set; }
 
         [Option(DefaultValue = ProxyType.System)]
         ProxyType Proxy { get; set; }